After fixing the error "document.querySelectorAll('.band2')[0] is not defined" that keeps appearing whenever viewing an advertisement (the flags thing to select the page language is not there when viewing an ad), the rest of the script started running and similar errors popped up; Added similar "only continue if xyz is found" fixes to the insert local/server time code and the logo insertion code;
Also did a clunky shift-around so that calls to actually run stuff are at the bottom.. Only did a quick hunt for stuff in try{}catch(){} blocks so might have missed other stuff but I think I got most of it.. The goal is to have a simple block at the end that calls all the functions so that it is much easier to see what is being run and in what order etc and make it easy to just comment out one buggy part if needs be so that the non-buggy functions can continue to run.

The long-term plan is to have a checkbox interface that can enable/disable parts of the script, basically 'modules'.